Fiddle about a bit, try to get inlining empty routines to work by increasing default inline size to 150 or 250 in case that was getting in the way.. unfortunately this doesn't seem to help, and it isn't clear at this time why a simple empty _init_ function is called when it should have been inlined..  argg .. oh yes, of course it is clear, the initialisation code is hand compiled in flxg, and never optimised :)

+++ src/gc/flx_gc.cpp
@@ -108,8 +108,8 @@ void *operator new(
 {
   if (amt != shape.amt * shape.count)
   {
-    fprintf(stderr,"Shape size error: allocator size = %ld\n",amt);
-    fprintf(stderr,"Shape %s element size = %ld, element count = %d\n",shape.cname,shape.amt,shape.count);
+    fprintf(stderr,"Shape size error: allocator size = %ld\n",(long)amt);
+    fprintf(stderr,"Shape %s element size = %ld, element count = %ld\n",shape.cname,(long)shape.amt,(long)shape.count);
     abort();
   }
   void *p = gcp.allocate(&shape,1,allow_gc); // dynamic array count = 1
